refactor(toast): move F0Toast to src/ui (internal) and drop src/internal
F0Toast is an internal-only component (used by ToastProvider, never exported),
so it belongs with the other internal building blocks under src/ui — the
existing "🔒 Internal" Storybook section (alongside ui/Dialog, ui/drawer,
ui/hover-card) — rather than in a bespoke top-level src/internal directory.

- Move internal/components/Toast/* -> ui/Toast/* and delete src/internal
  (including its README).
- Update imports to @/ui/Toast/*; story title "Components/Toast" -> "Toast"
  (renders as "🔒 Internal/Toast").
- Remove the now-unneeded src/internal Storybook story source.
