+++ packages/react/src/hooks/toast/__stories__/toasts.stories.tsx
@@ -3,11 +3,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite"
 import { useState } from "react"
 
 import { F0Button } from "@/components/F0Button"
-import {
-  F0ToastProps,
-  F0ToastVariant,
-  toastVariants,
-} from "@/internal/components/Toast/types"
+import { F0ToastProps, F0ToastVariant, toastVariants } from "@/ui/Toast/types"
 
 import { toasts } from "../imperative"
 import { ToastProvider } from "../ToastProvider"
